+++ src/components/Footer.js
@@ -1,0 +1,29 @@
+import React from 'react';
+
+import Uuid from './Uuid';
+
+
+const styles = {
+  border: '1px solid #808e9b',
+  backgroundColor: 'rgb(210 218 226 / 48%)',
+  padding: 8,
+  bottom: 0,
+  width: '100%',
+  textAlign: 'center',
+  p:{
+    color: '#1e272e',
+    fontSize: '10px',
+    fontWeight: 'bold',
+  },
+};
+
+function Footer () {
+  return(
+    <footer style={styles}>
+      <p style={styles.p}>Copy: Ela Wróbel 2020/2021</p>
+      <Uuid />
+    </footer>
+  );
+}
+
+export default Footer;
